+++ src/js/caroulix.js
@@ -32,6 +32,8 @@
     constructor(element, options, isLoadedWithData) {
       super();
 
+      Axentix.instances.push(this);
+
       this.el = document.querySelector(element);
 
       this.options = Axentix.getComponentOptions('Caroulix', options, this.el, isLoadedWithData);