+++ src/js/dropdown.js
@@ -21,6 +21,8 @@
     constructor(element, options, isLoadedWithData) {
       super();
 
+      Axentix.instances.push(this);
+
       this.el = document.querySelector(element);
 
       this.options = Axentix.getComponentOptions('Dropdown', options, this.el, isLoadedWithData);