+++ src/js/modal.js
@@ -21,6 +21,8 @@
     constructor(element, options, isLoadedWithData) {
       super();
 
+      Axentix.instances.push(this);
+
       this.el = document.querySelector(element);
 
       this.options = Axentix.getComponentOptions('Modal', options, this.el, isLoadedWithData);