+++ src/js/utilities.js
@@ -48,3 +48,48 @@ Axentix.createEvent = (element, eventName, extraData) => {
 Axentix.isTouchEnabled = () => {
   return 'ontouchstart' in window || navigator.maxTouchPoints > 0 || navigator.msMaxTouchPoints > 0;
 };
+
+/**
+ * Get instance of element
+ * @param {String} element Id of element
+ */
+Axentix.getInstance = (element) => {
+  return Axentix.instances.filter((instance) => '#' + instance.el.id === element)[0];
+};
+
+/**
+ * Get all instances
+ */
+Axentix.getAllInstances = () => {
+  return Axentix.instances;
+};
+
+/**
+ * Sync instance of element
+ * @param {String} element Id of element
+ */
+Axentix.sync = (element) => {
+  Axentix.getInstance(element).sync();
+};
+
+/**
+ * Sync all instances
+ */
+Axentix.syncAll = () => {
+  Axentix.instances.map((instance) => instance.sync());
+};
+
+/**
+ * Reset instance of element
+ * @param {String} element Id of element
+ */
+Axentix.reset = (element) => {
+  Axentix.getInstance(element).reset();
+};
+
+/**
+ * Reset all instances
+ */
+Axentix.resetAll = () => {
+  Axentix.instances.map((instance) => instance.reset());
+};